refactor(api): move organizationForAdminSerializer to src/team
